Updated +not-found.tsx
We're back GitHub! Got my Macbook back from repair yesterday and today we back to coding!

Today I wanted to fix the not found a little bit since I accidentally went to it. Took out the top bar (make it more smooth), added in text breaks (make this cleaner) and tried to add in better styling for the bottom button.

Going to keep working on this page

Thank you for reading!
.j.